+++ examples/drift_example/lib/screens/home.dart
@@ -10,7 +10,7 @@ import 'home/drawer.dart';
 import 'home/state.dart';
 
 class HomePage extends StatefulWidget {
-  const HomePage({Key? key}) : super(key: key);
+  const HomePage({super.key});
 
   @override
   State<HomePage> createState() => _HomePageState();