+++ packages/signals_core/lib/signals_core.dart
@@ -17,3 +17,4 @@ export 'src/value/value.dart';
 export 'src/core/container.dart';
 export 'src/core/signals.dart';
 export 'src/core/wrappers.dart';
+export 'src/devtool.dart';